+++ package.json
@@ -19,9 +19,10 @@
   "dependencies": {
     "body-parser": "1.14.1",
     "express": "4.13.3",
+    "express-session": "^1.14.1",
     "fs": "0.0.2",
-    "multer": "1.1.0",
-    "mongojs": "^0.18.1"
+    "mongojs": "^0.18.1",
+    "multer": "1.1.0"
   },
   "devDependencies": {
     "grunt": "^1.0.1",
